requirements.txt
langchain
langchain_community
google-generativeai
langchain-google-genai
faiss-cpu
fastapi[standard]
uvicorn
pydantic
python-dotenv
python-multipart
PyPDF2
unstructured
youtube_transcript_api
llama-index
pypdf
pymongo
pytest
groq
langchain-groq
jinja2
tiktoken
redis
python-jose
sendgrid
llama-index-llms-groq
llama-index-embeddings-langchain
slowapi
# grpcio
# crewai
# crewai-tools
# proto-plus
# langchain-cohere
